Add files via upload
This collection of Latex files can be used as a template for a doctoral dissertation at UNC Gillings School of Global Public Health. They are based on my own dissertation, which was accepted in April 2017. One important note is that the appendix format is not up to the grad school format standards. If you want to include an appendix, you may need to tweak a few of the settings.